This report, by its very length, defends itself against the risk of being read

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -11,9 +11,9 @@
  *
  * Return: The length of @str.
 */
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
-	int length = 0;
+	size_t length = 0;
 
 	while (*str++)
 		length++;
